shit werkt nog niet

+++ JobShop.h
@@ -25,9 +25,9 @@ public:
 private:
 	unsigned short amountOfJobs;
 	unsigned short amountOfMachines;
-	unsigned long currentTime;
+	unsigned short currentTime;
 	std::vector<Job> jobList;
-	std::map<unsigned long, unsigned short> times;
+	std::map<unsigned short, unsigned short> times;
 	std::map<unsigned short, bool> machines;
 	std::map<unsigned short, bool> jobsDone;
 };
